Bail out early during gimplify_asm_expr [PR121391]
In case an asm operand is an error node, constraints etc. are still
validated.  Furthermore, all other operands are gimplified, although an
error is returned in the end anyway.  For hard register constraints an
operand is required in order to determine the mode from which the number
of registers follows.  Therefore, instead of adding extra guards, bail
out early.

gcc/ChangeLog:

	PR middle-end/121391
	* gimplify.cc (gimplify_asm_expr): In case an asm operand is an
	error node, bail out early.

gcc/testsuite/ChangeLog:

	* gcc.dg/pr121391-1.c: New test.
	* gcc.dg/pr121391-2.c: New test.