ctf: Fix struct size truncation in 32-bit hosts [PR121903,PR121411]
The 'size' argument of ctf_add_sou was size_t.  After the prior fixes
for PR121411, this could cause the struct size to be truncated when
encoding extremely large structs on a host where size_t is smaller than
unsigned HOST_WIDE_INT, manifesting for example as the test failure
reported in PR121903.  Change the argument to uHWI to resolve the issue.

	PR debug/121411
	PR debug/121903

gcc/

	* ctfc.h (ctf_add_sou): Change size arg from size_t to uHWI.
	* ctfc.cc (ctf_add_sou): Likewise.