[ppc] adjust configure test for large TOC support
The use of the TLS register in a TOC/GOT address computation was
probably a cut&pasto or a thinko.  It causes a linker warning and,
because the TLS access in the test is incomplete, may cause
significant confusion.  Adjust to use the TOC/GOT register as base.


for  gcc/ChangeLog

	* configure.ac: Adjust base register in linker test for large
	TOC support.
	* configure: Rebuild.