i386/testsuite: Correct res_ref2 array size for avx512bw-vpmov{,us}wb-2.c
Both of the tests under 128 bit are raising:

warning: writing 16 bytes into a region of size 8 [-Wstringop-overflow=]

when compiling, leading to a test fail. The warning is caused by the
incorrect array size for res_ref2. The wrong size caused the overflow.

Correct them in this patch to fix the test fail.

gcc/testsuite/ChangeLog:

	* gcc.target/i386/avx512bw-vpmovuswb-2.c: Correct res_ref2
	array size.
	* gcc.target/i386/avx512bw-vpmovwb-2.c: Ditto.