c++, contracts: Abstract interfaces to constexpr [NFC].
We want to move to having different representations of the contract
semantic for C++26, since that wants values outside the range that
can be accommodated using the cxx2a mechanisms.  First part, abstract
the interfaces to constexpr, so that they do not see the underlying
source of data.

gcc/cp/ChangeLog:

	* constexpr.cc (cxx_eval_constant_expression): Use revised
	interfaces to determine if contracts are ignored and, if not,
	whether they are evaluated.
	* contracts.h (contract_ignored_p, contract_evaluated_p): New.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>