Cleanup vect_get_num_copies API
The following removes the dual non-SLP/SLP API in favor of only
handling SLP.  This also removes the possibility to override
vectype of a SLP node with an inconsistent one while still using
the SLP nodes number of lanes.  This requires adjustment of
a few places where such inconsistencies happened.

	* tree-vectorizer.h (vect_get_num_copies): Remove unused
	overload.  Remove defaulted vectype argument and always
	use the vector type of the SLP node.
	* tree-vect-loop.cc (vect_reduction_update_partial_vector_usage):
	Adjust.
	(vectorizable_lane_reducing): Likewise.
	(vect_transform_reduction): Likewise.
	(vectorizable_nonlinear_induction): Likewise.
	* tree-vect-stmts.cc (check_load_store_for_partial_vectors):
	Likewise.